Implement spinlock mechanism for process synchronization; update kernel code for multi-core support

+++ common/common.h
@@ -38,7 +38,7 @@ typedef int bool;
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long long uint64_t;
+typedef unsigned long uint64_t;
 typedef uint64_t size_t;
 typedef uint64_t paddr_t;
 typedef uint64_t vaddr_t;
